refactor: various fixes
